Separate archivers into their own packages
